pyproject.toml
[tool.poetry]
name = "tasker"
version = "0.1.0"
description = ""
authors = ["gglamer <azamat201760@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyjwt = "^2.8.0"
requests = "^2.31.0"
bs4 = "^0.0.1"
pydantic = "^2.4.2"
pyqt5 = "^5.15.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.292"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "W", "F", "Q"]
ignore = []

fixable = ["ALL"]
unfixable = []

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "ui",
]
per-file-ignores = {}

line-length = 123

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py311"

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"